worked on the transaction service and payment service on 05/07/2024

+++ data/repositories/walletRepository.go
@@ -3,15 +3,15 @@ package repositories
 import "eazyWallet/data/models"
 
 type WalletRepository interface {
-	BaseRepository[models.Account, uint]
+	BaseRepository[models.Account, uint64]
 }
 
 type NewEasyWalletRepositoryImpl struct {
-	*BaseRepositoryImpl[models.Account, uint]
+	*BaseRepositoryImpl[models.Account, uint64]
 }
 
 func NewWalletRepository() WalletRepository {
 	return &NewEasyWalletRepositoryImpl{
-		&BaseRepositoryImpl[models.Account, uint]{},
+		&BaseRepositoryImpl[models.Account, uint64]{},
 	}
 }
